rename: exception 공통 관리하도록 파일 이동
