+++ src/main/java/com/zerobase/munbanggu/auth/JwtAuthenticationFilter.java
@@ -3,7 +3,7 @@ package com.zerobase.munbanggu.auth;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zerobase.munbanggu.common.dto.ErrorResponse;
 import com.zerobase.munbanggu.common.type.ErrorCode;
-import com.zerobase.munbanggu.user.exception.InvalidTokenException;
+import com.zerobase.munbanggu.common.exception.InvalidTokenException;
 import java.io.IOException;
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
