--- src/main/java/com/zerobase/munbanggu/common/exception/ExceptionHandler.java
@@ -1,8 +1,0 @@
-package com.zerobase.munbanggu.common.exception;
-
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ExceptionHandler {
-
-}
