+++ src/main/java/com/zerobase/munbanggu/common/exception/InvalidTokenException.java
@@ -1,9 +1,10 @@
-package com.zerobase.munbanggu.user.exception;
+package com.zerobase.munbanggu.common.exception;
 
 import com.zerobase.munbanggu.common.type.ErrorCode;
 import io.jsonwebtoken.JwtException;
 import lombok.Getter;
 
+// 유효하지 않는 토큰으로 API 요청하는 경우 발생하는 예외
 @Getter
 public class InvalidTokenException extends JwtException {
