+++ src/main/java/com/zerobase/munbanggu/common/exception/NoPermissionException.java
@@ -1,8 +1,9 @@
-package com.zerobase.munbanggu.studyboard.exception;
+package com.zerobase.munbanggu.common.exception;
 
 import com.zerobase.munbanggu.common.type.ErrorCode;
 import lombok.Getter;
 
+// 스터디원 아닌 회원이 스터디에 접근하려는 경우 발생하는 예외
 @Getter
 public class NoPermissionException extends RuntimeException {
