+++ src/main/java/com/zerobase/munbanggu/common/exception/NotFoundStudyException.java
@@ -1,8 +1,9 @@
-package com.zerobase.munbanggu.study.exception;
+package com.zerobase.munbanggu.common.exception;
 
 import com.zerobase.munbanggu.common.type.ErrorCode;
 import lombok.Getter;
 
+// 존재하지 않는 스터디 조회하는 경우 발생하는 예외
 @Getter
 public class NotFoundStudyException extends RuntimeException {
