+++ src/main/java/com/zerobase/munbanggu/study/exception/StudyExceptionHandler.java
@@ -1,7 +1,7 @@
 package com.zerobase.munbanggu.study.exception;
 
 import com.zerobase.munbanggu.common.dto.ErrorResponse;
-import com.zerobase.munbanggu.user.exception.InvalidTokenException;
+import com.zerobase.munbanggu.common.exception.InvalidTokenException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -9,11 +9,6 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
 public class StudyExceptionHandler {
-
-    @ExceptionHandler
-    public ResponseEntity<ErrorResponse> notFoundStudyExceptionHandler(NotFoundStudyException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse(e.getErrorCode(), e.getMessage()));
-    }
 
     @ExceptionHandler(InvalidTokenException.class)
     public ResponseEntity<ErrorResponse> invalidTokenExceptionHandler(InvalidTokenException e) {
