+++ src/main/java/com/zerobase/munbanggu/studyboard/controller/StudyBoardController.java
@@ -4,11 +4,11 @@ import static com.zerobase.munbanggu.common.type.ErrorCode.INVALID_REQUEST_BODY;
 import static com.zerobase.munbanggu.common.type.ErrorCode.INVALID_TOKEN;
 
 import com.zerobase.munbanggu.common.dto.PageResponse;
-import com.zerobase.munbanggu.studyboard.exception.InvalidRequestBodyException;
+import com.zerobase.munbanggu.common.exception.InvalidRequestBodyException;
+import com.zerobase.munbanggu.common.exception.InvalidTokenException;
 import com.zerobase.munbanggu.studyboard.model.dto.PostRequest;
 import com.zerobase.munbanggu.studyboard.model.dto.PostResponse;
 import com.zerobase.munbanggu.studyboard.service.StudyBoardService;
-import com.zerobase.munbanggu.user.exception.InvalidTokenException;
 import java.util.HashMap;
 import java.util.Map;
 import javax.validation.Valid;
