+++ src/main/java/com/zerobase/munbanggu/studyboard/service/StudyBoardService.java
@@ -1,13 +1,15 @@
 package com.zerobase.munbanggu.studyboard.service;
 
 import com.zerobase.munbanggu.auth.TokenProvider;
-import com.zerobase.munbanggu.study.exception.NotFoundStudyException;
+import com.zerobase.munbanggu.common.exception.NoPermissionException;
+import com.zerobase.munbanggu.common.exception.NotFoundPostException;
+import com.zerobase.munbanggu.common.exception.NotFoundStudyException;
+import com.zerobase.munbanggu.common.exception.NotFoundUserException;
+import com.zerobase.munbanggu.common.type.ErrorCode;
 import com.zerobase.munbanggu.study.model.entity.Study;
 import com.zerobase.munbanggu.study.model.entity.StudyMember;
 import com.zerobase.munbanggu.study.repository.StudyMemberRepository;
 import com.zerobase.munbanggu.study.repository.StudyRepository;
-import com.zerobase.munbanggu.studyboard.exception.NoPermissionException;
-import com.zerobase.munbanggu.studyboard.exception.NotFoundPostException;
 import com.zerobase.munbanggu.studyboard.model.dto.PostRequest;
 import com.zerobase.munbanggu.studyboard.model.dto.PostResponse;
 import com.zerobase.munbanggu.studyboard.model.dto.VoteOptionRequest;
@@ -17,14 +19,11 @@ import com.zerobase.munbanggu.studyboard.model.entity.VoteOption;
 import com.zerobase.munbanggu.studyboard.repository.StudyBoardPostRepository;
 import com.zerobase.munbanggu.studyboard.repository.VoteRepository;
 import com.zerobase.munbanggu.studyboard.type.Type;
-import com.zerobase.munbanggu.common.type.ErrorCode;
-import com.zerobase.munbanggu.user.exception.NotFoundUserException;
 import com.zerobase.munbanggu.user.model.entity.User;
 import com.zerobase.munbanggu.user.repository.UserRepository;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
