+++ src/main/java/com/zerobase/munbanggu/studyboard/service/StudyCommentService.java
@@ -1,16 +1,16 @@
 package com.zerobase.munbanggu.studyboard.service;
 
 import com.zerobase.munbanggu.auth.TokenProvider;
-import com.zerobase.munbanggu.studyboard.exception.NoPermissionException;
-import com.zerobase.munbanggu.studyboard.exception.NotFoundPostException;
+import com.zerobase.munbanggu.common.exception.NoPermissionException;
+import com.zerobase.munbanggu.common.exception.NotFoundPostException;
+import com.zerobase.munbanggu.common.exception.NotFoundUserException;
+import com.zerobase.munbanggu.common.type.ErrorCode;
 import com.zerobase.munbanggu.studyboard.model.dto.CommentRequest;
 import com.zerobase.munbanggu.studyboard.model.dto.CommentResponse;
-import com.zerobase.munbanggu.studyboard.model.entity.StudyComment;
 import com.zerobase.munbanggu.studyboard.model.entity.StudyBoardPost;
-import com.zerobase.munbanggu.studyboard.repository.StudyCommentRepository;
+import com.zerobase.munbanggu.studyboard.model.entity.StudyComment;
 import com.zerobase.munbanggu.studyboard.repository.StudyBoardPostRepository;
-import com.zerobase.munbanggu.common.type.ErrorCode;
-import com.zerobase.munbanggu.user.exception.NotFoundUserException;
+import com.zerobase.munbanggu.studyboard.repository.StudyCommentRepository;
 import com.zerobase.munbanggu.user.model.entity.User;
 import com.zerobase.munbanggu.user.repository.UserRepository;
 import java.util.List;
